refactor: ApplicationQueryClient
